Compressed all models back to a single file
- Far too many ciruclar dependency issues I kept running into.
