+++ pySpaceTraders/constants.py
@@ -1,6 +1,8 @@
 __version__ = "0.5.0"
 
-from typing import List, Literal
+from typing import List
 
-V2_STARTRADERS_URL: str = "https://api.spacetraders.io/v2"
+V2_STARTRADERS_URL: str = (
+    "https://stoplight.io/mocks/spacetraders/spacetraders/96627693"  # "https://api.spacetraders.io/v2"
+)
 REQUEST_TYPES: List[str] = ["GET", "POST", "PATCH"]
