--- pySpaceTraders/models/constructionsites.py
@@ -1,18 +1,0 @@
-from dataclasses import dataclass
-from typing import List
-
-from pySpaceTraders.models.enums import TradeSymbol
-
-
-@dataclass
-class ConstructionMaterial:
-    tradeSymbol: TradeSymbol
-    required: int
-    fulfilled: int
-
-
-@dataclass
-class ConstructionSite:
-    symbol: str
-    materials: List[ConstructionMaterial]
-    isComplete: bool
