--- pySpaceTraders/models/jumpgates.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-from typing import List
-
-
-@dataclass
-class JumpGate:
-    symbol: str
-    connection: List[str] | None
